feat: implement functionality for mousedown events
make keys react properly on mouse inputs
handle focus state of textarea
