+++ virtual-keyboard/src/components/keyboardKey/keyboardKey.scss
@@ -11,17 +11,30 @@
   cursor: pointer;
   transition: all 0.2s;
 
+  & span {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+  }
+
   &:hover {
-    background-color: #da7f77;
+    background-color: #b84136;
+  }
+
+  &:active {
+    background-color: #c52618;
+    border-radius: 10px;
   }
 
   &_active {
     background-color: #c52618;
-    border-radius: 6px;
+    border-radius: 10px;
   }
 }
 
-.hidden {
+.hidden, span.hidden {
   display: none;
 }
